fix: Move time functions into file
